[FIX] 모달 뷰를 virtualPage 뷰로 수정, name 입력시 path 만들어지게 수정
