+++ src/components/AnalyticsProvider/index.tsx
@@ -5,7 +5,7 @@ import {UnknownRecord} from '../../types/common';
 
 interface Props {
   onInitialize(): void;
-  onModalView?(name: string, params?: UnknownRecord): void;
+  onVirtualPageView?(name: string, params?: UnknownRecord): void;
   onPageView?(params?: UnknownRecord): void;
   onEvent?(name: string, params?: UnknownRecord): void;
   onClick?(name: string, params?: UnknownRecord): void;
@@ -14,7 +14,7 @@ interface Props {
 
 export function AnalyticsProvider({
   onInitialize,
-  onModalView = () => null,
+  onVirtualPageView = () => null,
   onPageView = () => null,
   onEvent = () => null,
   onClick = () => null,
@@ -28,7 +28,7 @@ export function AnalyticsProvider({
     () => (
       <AnalyticsProviderContext.Provider
         value={{
-          onModalView,
+          onVirtualPageView,
           onPageView,
           onEvent,
           onClick,
@@ -37,6 +37,6 @@ export function AnalyticsProvider({
         {children}
       </AnalyticsProviderContext.Provider>
     ),
-    [children, onClick, onEvent, onPageView, onModalView],
+    [children, onClick, onEvent, onPageView, onVirtualPageView],
   );
 }
