+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export * from './AnalyticsProvider';
 export * from './AnalyticsPageView';
-export * from './AnalyticsModalView';
+export * from './AnalyticsVirtualPageView';
