+++ src/contexts/AnalyticsProviderContext.ts
@@ -2,14 +2,14 @@ import {createContext} from 'react';
 import {UnknownRecord} from '../types/common';
 
 export interface AnalyticsProviderContext {
-  onModalView(name: string, params?: UnknownRecord): void;
+  onVirtualPageView(name: string, params?: UnknownRecord): void;
   onPageView(params?: UnknownRecord): void;
   onEvent(name: string, params?: UnknownRecord): void;
   onClick(name: string, params?: UnknownRecord): void;
 }
 
 export const initialState: AnalyticsProviderContext = {
-  onModalView: () => null,
+  onVirtualPageView: () => null,
   onPageView: () => null,
   onEvent: () => null,
   onClick: () => null,
